buy credit

+++ lib/app/domain/entities/plans/plan_entity.dart
@@ -1,7 +1,7 @@
 class PlanEntity {
   String? id;
   String? name;
-  double? credit;
+  int? credit;
   double? value;
 
   PlanEntity({
@@ -14,11 +14,11 @@ class PlanEntity {
   PlanEntity.fromJson(Map<String, dynamic> parsedJson) {
     id = parsedJson['_id'] == null ? null : parsedJson['_id'] as String;
     name = parsedJson['name'] == null ? null : parsedJson['name'] as String;
-    credit = parsedJson['credit'] == null ? null : parsedJson['credit'] as double;
+    credit = parsedJson['credit'] == null ? null : parsedJson['credit'] as int;
     value = parsedJson['value'] == null ? null : parsedJson['value'] as double;
   }
 
   String get nameStr => name ?? '';
-  double get creditVl => credit ?? 0;
+  int get creditVl => credit ?? 0;
   double get valueVl => value ?? 0;
 }
